sql: move RandString to util to break coldata->rangden->sem/tree dep
Epic: CRDB-18892
Informs: #91831
Release note: None
