+++ pkg/sql/randgen/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "expr.go",
         "mutator.go",
         "schema.go",
-        "string.go",
         "type.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/randgen",
@@ -39,6 +38,7 @@ go_library(
         "//pkg/sql/sem/volatility",
         "//pkg/sql/stats",
         "//pkg/sql/types",
+        "//pkg/util",
         "//pkg/util/bitarray",
         "//pkg/util/duration",
         "//pkg/util/encoding",
