+++ pkg/sql/randgen/schema.go
@@ -27,6 +27,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
+	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 	"github.com/cockroachdb/errors"
 	"github.com/lib/pq/oid"
@@ -53,7 +54,7 @@ func RandCreateType(rng *rand.Rand, name, alphabet string) tree.Statement {
 	labelsMap := make(map[string]struct{})
 	i := 0
 	for i < numLabels {
-		s := RandString(rng, rng.Intn(6)+1, alphabet)
+		s := util.RandString(rng, rng.Intn(6)+1, alphabet)
 		if _, ok := labelsMap[s]; !ok {
 			labels[i] = tree.EnumValue(s)
 			labelsMap[s] = struct{}{}
